service-workers/service-worker/clients-matchall-on-evaluation.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.matchAll on script evaluation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
  t.add_cleanup(async () => {
    if (registration) await registration.unregister();
  });

  const SCRIPT = 'resources/clients-matchall-on-evaluation-worker.js';
  const SCOPE = 'resources/blank.html?clients-matchAll-on-evaluation';

  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  const saw_message = new Promise(function(resolve) {
    navigator.serviceWorker.onmessage = function(e) {
      resolve(e.data);
    };
  });

  const results = await saw_message;
  assert_equals(results, 'matched');
}, 'Test Clients.matchAll() on script evaluation');
</script>
